requirements.txt
fastapi[all]
uvicorn
pydantic
requests
python-multipart
groq
python-dotenv